requirements.txt
asgiref==3.8.1
attrs==24.2.0
CacheControl==0.14.0
cachetools==5.5.0
cattrs==24.1.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
cryptography==43.0.1
Django==5.1.1
djangorestframework==3.15.2
drf-spectacular==0.27.2
flatbuffers==24.3.25
google-api-core==2.19.2
google-api-python-client==2.145.0
google-auth==2.34.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.18.0
google-cloud-storage==2.18.2
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.65.0
grpcio==1.66.1
grpcio-status==1.66.1
httplib2==0.22.0
idna==3.9
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
msgpack==1.1.0
numpy==2.1.1
openmeteo_requests==1.3.0
openmeteo_sdk==1.15.0
pandas==2.2.2
platformdirs==4.3.3
proto-plus==1.24.0
protobuf==5.28.1
psycopg2==2.9.9
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyJWT==2.9.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
requests-cache==1.2.1
retry-requests==2.0.0
rpds-py==0.20.0
rsa==4.9
six==1.16.0
sqlparse==0.5.1
tzdata==2024.1
uritemplate==4.1.1
url-normalize==1.4.3
urllib3==2.2.3
